requirements.txt
google-api-python-client==2.126.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
pandas==2.2.2
pydantic==2.7.0
python-dotenv==1.0.1